requirements.txt
nextcord @ git+https://github.com/Renaud11232/nextcord.git@01c5625ba7180c74f75633d4841712273366b833
PyNaCl==1.4.0
python-dotenv==1.1.1
yt-dlp==2025.7.21